refactor(view): プラン詳細に関わるコンポーネントを移動
