+++ src/stories/plandetail/PlanPlaceList.stories.tsx
@@ -1,9 +1,9 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { mockPlan } from "src/stories/mock/plan";
-import { PlanPlaceList } from "src/view/plan/PlanPlaceList";
+import { PlanPlaceList } from "src/view/plandetail/PlanPlaceList";
 
 export default {
-    title: "plan/PlanPlaceList",
+    title: "plan_detail/PlanPlaceList",
     component: PlanPlaceList,
 } as ComponentMeta<typeof PlanPlaceList>;
